[JII Texts][US_IX] Include new logic to select phone numbers (Recidiviz/recidiviz-data#39003)
## Description of the change

This PR pulls phone numbers from raw data tables in Idaho such that we
first try to select a person's primary _cell_ number before defaulting
to the logic that hydrates `state_person`. See
[thread](https://recidiviz.slack.com/archives/C05GR9GA9B3/p1741052340795819)
for more context

Note: I loaded this to `alexa_jii_texting` and also ran the
`compare_views.py` script, but there are no differences in the
comparison output
[here](https://console.cloud.google.com/bigquery?ws=!1m4!1m3!3m2!1srecidiviz-staging!2salexa__jii_texting_comparison_output)
against staging data. @samanthahuff @santymendoza could this be a
staging vs. prod issue? or maybe we somehow already are capturing the
logic that thomas is hoping to achieve?

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: ef9a323fa1361b42601d43415069bcc5903d18f0