[US_AZ] Filter invalid sentence dates (Recidiviz/recidiviz-data#39027)
## Description of the change

I found some `offense_date` values that are outside of the acceptable
range after validations added in Recidiviz/recidiviz-data#38809. It seems like there is an
exemption for AZ for this validation, but for some reason it still
caused an unrelated ingest pipeline to fail (see
https://github.com/Recidiviz/recidiviz-data/pull/39021 and
[pipeline](https://console.cloud.google.com/dataflow/jobs/us-east1/2025-03-04_14_22_42-17274686082236909271;graphView=1?project=recidiviz-staging&inv=1&invt=AbrKoA&pageState=(%22dfTime%22:(%22l%22:%22dfJobMaxTime%22))))
.

This filters out `offense_date` values that fall outside of the
acceptable range at ingest time. Instead of filtering the charges with
odd offense dates out of the results entirely, we replace the bad dates
with a NULL value instead. This is based on the assumption that since
some offenses are disqualifying for certain opportunities, we want to
know that a person was convicted of a given offense even if we don't
know when it occurred.

[sandbox pipeline
succeeded](https://console.cloud.google.com/dataflow/jobs/us-east1/2025-03-04_15_10_31-11053868320975723894;graphView=1?project=recidiviz-staging&pageState=(%22dfTime%22:(%22l%22:%22dfJobMaxTime%22)))
## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 544505c8fa4485ec153e3c191b6f1b3cae13af2d