Add field with metric month as string to breadth and depth LookML generator to facilitate intuitive dashboard filtering by month (Recidiviz/recidiviz-data#38701)
## Description of the change

We want people to be able to filter the 2025 Breadth and Depth Metrics
Dashboard to a specific month or span of months. Since the
`metric_month` field is a Date field, adding a `field_filter` to the
dashboard on that field gives the dashboard user a date picker for that
filter.

The problem with a date picker is that the user can choose any days they
want, but the dashboard is only filterable to whole months at a time,
since we only calculate baselines for whole months. The user experience
with a date picker filter means they could select a time frame like
`01/15/25` to `02/15/25`, but the resulting data would span `01/01/25`
to `02/28/25`, which is confusing.

I have not found a way to restrict the granularity of options for a Date
`field_filter` to entire months only (please let me know if you know of
one!). It seems like you should be able to use a parameter, but you
cannot use a multi-select filter with a parameter.

Instead, here I've added a truncated string version of the
`metric_month` field to allow using a string filter where the only
options will be year/month like 2025-01, 2024-12, 2024-11, etc.

This PR is associated with the following Looker PR that adds the breadth
and depth metrics dashboard:
https://github.com/Recidiviz/looker/pull/704

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Advances https://github.com/Recidiviz/looker/issues/665

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 6dce3dbbe62d67d836922087a0b20a363121aa78