[US_AZ] Filter out less recent signature form statuses (Recidiviz/recidiviz-data#39065)
## Description of the change

We are seeing a bug in the TPR / DTP tool in AZ where some people who
have signed their agreement forms for release are showing as not having
signed. This is because we are tracking every update to the signature
form status and attaching them to the same incarceration session, but
not selecting only the most recent update anywhere in the logic before
we display in the tool. So when a person had previously not signed and
later did sign, we are nondeterministically selecting which of those
statuses to show in the tool.

This change filters the results of the `us_az_agreement_form_signatures`
view to only contain the most recent signature form status for each
person and incarceration stint. I loaded this view to a sandbox and used
this query to confirm there is only one signature status per person +
incarceration stint + program:

```
SELECT
  person_id,
  doc_id,
  program,
  COUNT(DISTINCT sig_status) AS n
FROM
  recidiviz-staging.elise_analyst_data.us_az_agreement_form_signatures
GROUP BY
  1,  2,  3
HAVING
  n > 1
```

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: c7121479335c5ddc74a8423e12dd3392c21952d0