Migrate supervision past half full term release date criteria to v2 infra (Recidiviz/recidiviz-data#39112)
## Description of the change

This PR migrates the
`SUPERVISION_OR_SUPERVISION_OUT_OF_STATE_PAST_HALF_FULL_TERM_RELEASE_DATE`
criteria to sentencing v2 infra. This criteria migration is prioritized
because it unblocks UT ET.

The only downstream live opportunity that this criteria impacts is
`usMiEarlyDischarge` which is through the tasks
`complete_discharge_early_from_parole_dual_supervision_request` and
`complete_discharge_early_from_probation_supervision_request`. This
[notebook](https://colab.research.google.com/drive/1DV68vrvCi9pho7jpSOa1VuxZ3YAt-2Ei?usp=sharing)
shows the validation to confirm no changes in eligible population for
this opportunity.

A couple things to note with this migration:
1. I dropped use of the `join_sentence_spans_to_compartment_sessions`
because it does not appear to be necessary and simplifies the migration
(will cause some changes at the criteria level, but redundant with
exclusion made by candidate population)
2. I did need to hack in another reference to `sentences_preprocesed`
for v1 states to avoid changes to the eligible population for MI early
discharge. This is due to differences between the sentence
`effective_date` and `date_imposed` for MI probation sentences. For V1
states, we are constructing serving periods based on `date_imposed`,
which means in the v2 infra that we don't really have a way of accessing
the `effective_date` for v1 states. The `effective_date` is used for the
calculation of half past max term. I created a new view
`sentence_first_serving_date` which for v1 states is the effective_date
from `sentences_preprocessed`, and for v2 states is the min start_date
from `sentence_serving_periods`. I believe that as states get migrated
to v2 infra, the min serving date should work in all cases, which will
eventually allow us to deprecate the dependency on
`sentences_preprocessed`. Definitely open to suggestions on other
approaches here!

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#39113

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 45e0991f935eac7c30238dbeb2bc652dd2f47ea9