[Auth] Add script to delete users blocked 30+ days (Recidiviz/recidiviz-data#38950)
## Description of the change

This PR adds a script that will identify and delete users who have been
blocked for at least 30 days and are not in Roster. The script will
delete users from the UserOverride and PermissionsOverride tables. It
filters users by state code and prompts for confirmation before
deleting.

I tested this with my local dev db and it works, but I wasn't able to
test it with the Cloud SQL db, which might just be a "me" permissions
issue? Or maybe I'm missing a step, I'm not sure. In any case, if you
could look at that in particular that would be great!

```
google.api_core.exceptions.PermissionDenied: 403 Permission 'secretmanager.versions.access' denied for resource 'projects/recidiviz-staging/secrets/case_triage_cloudsql_instance_id/versions/latest' (or it may not exist).
```

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Part of Recidiviz/recidiviz-data#36604

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [X] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [X] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: c67281ec3727d9b0aba675f2a6b873a9fc2e4a1d