[US_TX] Adds logic to SP view to account for multiple SID_Numbers for a single SP (Recidiviz/recidiviz-data#39122)
## Description of the change

There are a few, ~25 periods, associated with more than one SID_Number.
In these cases, the periods are usually associated to two ClientData
records with the same person info but different SID numbers, or in some
cases one of the extraneous SID numbers isn't even found in the client
data. This PR ranks those SID_Numbers associated with a given
Period_ID_Number, taking into consideration whether that SID_Number is
in the ClientData file and when the SID_Number was last seen, and uses
that rank to select a single SID for each Period ID number.

SUCCESSFUL sandbox here:
https://console.cloud.google.com/dataflow/jobs/us-east1/2025-03-05_16_10_39-13097025197045739650;graphView=1?project=recidiviz-staging&pageState=(%22dfTime%22:(%22l%22:%22dfJobMaxTime%22))
## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 52206f067fe9f3e43394cd58769f32e37c1dc906