[Workflows][US_IX] Add crcFaclityId location_record rows for CRC Facilities (Recidiviz/recidiviz-data#39182)
## Description of the change

Adds 6 new records to `location_record` with `id_type = crcFacilityId`
for all CRC Facilities in ID. This will allow us to search for CRC
Facilities to populate the search bar for ID Workflows.

This has been run in the `jovergaag_us_ix_locations` sandbox.

This will be followed by this [FE
PR](https://github.com/Recidiviz/recidiviz-dashboards/pull/7582) which
adjusts the location search to be by `crcFacilityId`.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Part of
#[7284](https://github.com/Recidiviz/recidiviz-dashboards/issues/7284)

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 565af8e324e0428f03e53a27f65b2d7241b4aadd