[Auth] Fix auth test timezone conversion (Recidiviz/recidiviz-data#39215)
## Description of the change

The affected tests freeze time in the local timezone, but the expected
`blockedOn` values for the tests were being generated in UTC directly,
which was causing the expected values to be off by one hour due to not
accounting for Daylight Saving Time. Even though DST doesn't go into
effect until Sunday morning, the tests are working with future dates
when DST will be in effect, so the timestamps are returned as if it is.

I adjusted how the expected values are calculated by generating them in
the local timezone first, then converting them to UTC, which accounts
for DST changes.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#39213

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [X] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [X] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 6bdc16d428fee0cdbc694cca6953275a3b969c57