[US_TX][sftp] delete files after downloading them (Recidiviz/recidiviz-data#38918)
## Description of the change

this PR adds the concept of the "post download action" -- a step that is
taken directly after we have successfully downloaded a file from SFTP.

n.b. i opted to add this directly after we download the file, instead of
after we upload the file successfully or mark it as downloaded
successfully in order for the change to have the smallest impact
surface. there is some argument to having this occur after we have
successfully uploaded the file to the ingest bucket instead, but i sort
of figured that any issue that occurs w file corruption wont be caught
until raw data import anyways which would necessarily require us to put
some sort of TTL on the file after download, which felt like much more
work than i wanted to do for this.

that being said, i think another lightweight option here would be adding
a mapped task after the post processing step; my main reason for not
doing this is it would map a no-op for all other files which felt a
little unnecessary

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#38448

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 37683ea3d02962cdd9d836e4dfbb87203c18f392