[Events] Add `previous_change_type` and `custody_level` to `custody_level_change` and `incarceration_incident` events (Recidiviz/recidiviz-data#39081)
## Description of the change

In order to enable backstop metrics for US_IX custody level downgrade as
outlined
[here](https://docs.google.com/document/d/1SyXex8LyotMUfQGS0sT8kOE6Qlfhjqr2QiJ5cK_5U7g/edit?tab=t.0).
We are adding the attributes
`previous_change_type` and `custody_level` to `custody_level_change` and
`incarceration_incident` events respectively. This change will allow us
to observe:
- custody level upgrades following downgrades within the same
compartment level 0 super session (this rate should stay the same or go
down)
- incarceration incidents by custody level. Since incidents vary at
custody levels, it will be useful to break down rates by custody level
to further understand why the overall rate is changing, if it does.

Tested in `shuff`

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: d90f091fc9cbb347c1b37121dbcef90a991dad33